requirements.txt
Flask==1.0.2
luma.led_matrix
tinydb
event-bus